Split filename generation logic out of RecorderThread
